test: the u32 oracle does its arithmetic in u32
test_ints' u32 Add/Sub/Mul oracles computed in int64_t and truncated
afterwards, so u32 Mul's own boundary pair -- 4294967295 squared -- was
signed overflow, undefined behavior in the oracle rather than in anything
vmlib does. The sanitizers job caught it and master has been red since.

Cast first and let the unsigned type wrap, which is what these three are
asserting in the first place, and what u32 Div/Mod two lines down already
do. Clean under -fsanitize=address,undefined.

+++ test/test_ints.cc
@@ -424,19 +424,22 @@ int main() {
   test_binop(
       BinOp::Add, "u32 Add", u32_pairs(false),
       [](int64_t a, int64_t bb) {
-        return static_cast<int64_t>(static_cast<uint32_t>(a + bb));
+        return static_cast<int64_t>(static_cast<uint32_t>(a) +
+                                    static_cast<uint32_t>(bb));
       },
       UnOp::WrapU32);
   test_binop(
       BinOp::Sub, "u32 Sub", u32_pairs(false),
       [](int64_t a, int64_t bb) {
-        return static_cast<int64_t>(static_cast<uint32_t>(a - bb));
+        return static_cast<int64_t>(static_cast<uint32_t>(a) -
+                                    static_cast<uint32_t>(bb));
       },
       UnOp::WrapU32);
   test_binop(
       BinOp::Mul, "u32 Mul", u32_pairs(false),
       [](int64_t a, int64_t bb) {
-        return static_cast<int64_t>(static_cast<uint32_t>(a * bb));
+        return static_cast<int64_t>(static_cast<uint32_t>(a) *
+                                    static_cast<uint32_t>(bb));
       },
       UnOp::WrapU32);
   test_binop(BinOp::Div, "u32 Div", u32_pairs(true),
